+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 0.2.0
 - Added CharacterData#stars and LightConeData#stars.
+- Added LightConeData#superimpositions.
 # 0.1.0
 - Released the first version.